requirements.txt
# Load .env
python-dotenv==1.0.0

# OpenAI client (v0.28.0)
openai==0.28.0

# RAG helper
langchain==0.0.267

# PDF parsing
PyPDF2==3.0.1

# Text processing
nltk==3.8.1
transformers==4.30.2

# Vector indexing & search
faiss-cpu==1.7.4
whoosh==2.7.4
scikit-learn==1.2.2

# Utilities
tqdm==4.65.0

# Additional dependencies
numpy==1.24.3
requests==2.31.0
torch==2.0.1